Replaced Block enum with BlockType and Block classes, allowing blocks to be defined at run time.
BlockType holds all constant state which is shared among all blocks of
the same type. For now, this is indices into the texture array. Block
holds all per-block information. At the moment, this is only a pointer to
BlockType, but this could eventually also hold things like orientation
or a pointer to some block-specific larger structure.