Replaced Block::Pos with glm::ivec3. Moved Face to util.h, and moved functions involving Face and Block::Pos to util.{h,cpp}.
Several of the functions that involved Face are applicable beyond
positions in chunks. In particular, they'll appear in code that deals
with chunks' positions in the ChunkGrid. Once these were seperate from
Chunk::Pos, Chunk::Pos was a glorified glm::ivec3, so eliminating it made
these functions usable in the same fashion while still being general enough
to apply to the future ChunkGrid.