Added ChunkGrid, which can hold multiple chunks positioned around the player.
Lost picking functionality in the process, as now we need to
handle picking between chunks, but the old picking was a hack
anyway.

Sometime soon some decisions will need to be made with regards to
holding positions in the world and in the chunk. For now, everything
is glm::vecx, which is internally floats. Those are good for the GPU,
but for handling large worlds may give precision issues. We probably
want glm::vecxd for game logic to avoid precision issues at moderate
sized worlds (10k-100k). However, that won't help rendering issues
which should also occur in that range, though they probably aren't
obvious until the next order of magnitude (100k-1M).