Wrote a custom Optional template.
While I was in a templatey-mood, I realized that by replacing
boost::optional I could potentially have no dependency on boost at
all. I might branch off a library for this purpose, as boost is massive,
and as of C++11 I only use a few classes from it, which aren't hard to
write.